+++ internal/convert/kayenta.go
@@ -21,6 +21,7 @@ import (
 	"github.com/spinnaker/kleat/api/client/config"
 )
 
+// HalToKayenta generates the kayanta config for the supplied config.Hal h.
 func HalToKayenta(h *config.Hal) *config.Kayenta {
 	if !h.GetCanary().GetEnabled() {
 		return &config.Kayenta{}
